Factor out AR and ER from UT
