+++ inc/LX.hpp
@@ -10,6 +10,7 @@
  *\INCLUDES
  *-----------------------------------------------------------------------------*/
 
+#include "ER.hpp"
 #include "UT.hpp"
 #include <variant>
